# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "YouTube Companion Dashboard API",
    "version": "1.0.0",
    "description": "API for YouTube Companion Dashboard that allows users to manage their YouTube videos, comments, and notes."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/notes/{videoId}": {
      "get": {
        "summary": "Get notes for a video",
        "description": "Get all notes for the specified video",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of notes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string" },
                      "userId": { "type": "string" },
                      "videoId": { "type": "string" },
                      "content": { "type": "string" },
                      "createdAt": { "type": "string", "format": "date-time" },
                      "updatedAt": { "type": "string", "format": "date-time" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Error fetching notes"
          }
        }
      }
    },
    "/notes": {
      "post": {
        "summary": "Add a note",
        "description": "Add a new note for a video",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "videoId": { "type": "string" },
                  "content": { "type": "string" }
                },
                "required": ["videoId", "content"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Note created successfully"
          },
          "400": {
            "description": "VideoId and content are required"
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Error creating note"
          }
        }
      }
    },
    "/notes/{id}": {
      "patch": {
        "summary": "Update a note",
        "description": "Update the content of an existing note",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": { "type": "string" }
                },
                "required": ["content"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Note updated successfully"
          },
          "400": {
            "description": "Content is required"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Note not found or unauthorized"
          },
          "500": {
            "description": "Error updating note"
          }
        }
      },
      "delete": {
        "summary": "Delete a note",
        "description": "Delete an existing note",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note deleted successfully"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Note not found or unauthorized"
          },
          "500": {
            "description": "Error deleting note"
          }
        }
      }
    },
...
  }
}